Rename invoice routes to documents
